Lazify `hyperspy.learn` module
