+++ hyperspy/learn/__init__.py
@@ -15,3 +15,44 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with HyperSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
+
+
+import importlib
+
+# ruff: noqa: F822
+
+__all__ = [
+    "mlpca",
+    "ornmf",
+    "orthomax",
+    "orpca",
+    "rpca_godec",
+    "svd_pca",
+    "svd_solve",
+    "whiten_data",
+]
+
+
+# mapping following the pattern: from value import key
+_import_mapping = {
+    "mlpca": "_mlpca",
+    "ornmf": "_ornmf",
+    "orthomax": "_orthomax",
+    "orpca": "_rpca",
+    "rpca_godec": "_rpca",
+    "svd_pca": "_svd_pca",
+    "svd_solve": "_svd_pca",
+    "whiten_data": "_whitening",
+}
+
+
+def __dir__():
+    return sorted(__all__)
+
+
+def __getattr__(name):
+    if name in __all__:
+        import_path = "hyperspy.learn." + _import_mapping.get(name)
+        return getattr(importlib.import_module(import_path), name)
+
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
