+++ hyperspy/learn/_orthomax.py
