+++ hyperspy/tests/learn/test_mlpca.py
@@ -19,7 +19,7 @@
 import numpy as np
 import pytest
 
-from hyperspy.learn.mlpca import mlpca
+from hyperspy.learn._mlpca import mlpca
 from hyperspy.signals import Signal1D
