+++ hyperspy/utils/peakfinders2D.py
@@ -17,6 +17,7 @@
 # along with HyperSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
 
 import copy
+import importlib
 
 import numpy as np
 import scipy.ndimage as ndi
@@ -24,11 +25,7 @@ from skimage.feature import blob_dog, blob_log, match_template, peak_local_max
 
 from hyperspy.decorators import jit_ifnumba
 
-try:
-    import sklearn
-except ImportError:
-    sklearn = None
-
+SKLEARN_INSTALLED = importlib.util.find_spec("sklearn") is not None
 
 NO_PEAKS = np.array([[np.nan, np.nan]])
 
@@ -384,7 +381,7 @@ def find_peaks_stat(z, alpha=1.0, window_radius=10, convergence_ratio=0.05):
     8. Repeat #4-7 until the number of peaks found in the previous step
        converges to within the user defined convergence_ratio.
     """
-    if sklearn is None:
+    if not SKLEARN_INSTALLED:
         raise ImportError("This method requires scikit-learn.")
 
     def normalize(image):
@@ -435,6 +432,8 @@ def find_peaks_stat(z, alpha=1.0, window_radius=10, convergence_ratio=0.05):
 
     def separate_peaks(binarised_image):
         """Identify adjacent 'on' coordinates via DBSCAN."""
+        import sklearn
+
         bi = binarised_image.astype("bool")
         coordinates = np.indices(bi.shape).reshape(2, -1).T[bi.flatten()]
         db = sklearn.cluster.DBSCAN(2, min_samples=3)
